SECOAUTH-221 - Support for registering multiple redirect_uri Change to the ClientDetails to allow for a Set<String> of redirect URIs. This is as per OAuth2 v2-23.
Whilst a seemingly straightforward change, this has some major implications to the JdbcClientDetailsService, as this now requires two separate SQL statements to set (when using a custom schema). It also requires the schema to include a new table to accommodate the 1-to-* relationship.

 If a client has one or more registered URIs, the client MUST specify the redirect_uri when requesting an OAuth token. If this is not present an OAuth2Exception will be thrown with the error message "A redirect_uri must be supplied.".

 If the client has no registered URIs, the behaviour is unchanged.

https://jira.springsource.org/browse/SECOAUTH-221

SECOAUTH-221 - Support for registering multiple redirect_uri
Revert the schema for JdbcClientDetailsService to follow the same pattern as scopes. Now uses CSVs saved to a VARCHAR(256)
https://jira.springsource.org/browse/SECOAUTH-221

SECOAUTH-221 - Support for registering multiple redirect_uri
Update to include multiple values in the redirect uri column.
https://jira.springsource.org/browse/SECOAUTH-221