Create SubdomainRedirectResolver
there are cases where the redirect-uri consists of a subdomain, and
each customer has its own subdomain, hence a different redirect-uri.
So we need a resolver that will validate the URL not as the default
impl (that checks startsWith() ), but something that will check only
the suffix of the context path.