Ensure even if token request asks for more scopes they are not granted
Only approved scopes should make it into the access token, but this constraint
was being broken by the AuthorizationCodeTokenGranter, inadvertently because
of the implementation of OAuth2Request.createOAuth2Request(). Since that method
is actually only called from one place it seems safe to just change it so that
only the explicitly approved scopes are in the result.

Fixes gh-151