More care with proxies in @Configuration
Apparently the rule of thumb is: if a @Bean uses a configurer
to grab one of it's dependencies then it should be a lazy proxy
in case it is used before the configurer has initialized.

In addition we find that an Application.class that extends
OAuth2AuthorizationServerConfigurerAdapter directly is problematic,
but if you nest it it works fine. Problems manifest themselves in
objects that ought to be proxies not being proxied (i.e. early
instantiation going on somewhere). This has no impact on
Spring OAuth (yet), but it affects applications that use it.