Remove parent directory path traversals in DefaultRedirectResolver
To prevent redirect matches in sibling and ancestor resources we need to
flatten the path when matching redirects. It's not a major vulnerability
especially since most registrations are in the form of a root resource,
but worth patching anyway. A similar bug contributed to a publicly revealed
hack on Github, but in the workds of the attacker "is not worth much on its
own".

Fixes gh-155