Delegate to DefaultResponseErrorHandler in OAuth2ContextSetup
OAuth2 requests in tests can fail in a number of ways, and the reason
a custom error handler is used in the OAuth2ContextSetup is so that
errors can be asserted (e.g. from status codes) and not just
blow up in the user's face. The problem with the old implementation
was that the custom error handler never threw an exception, so it
would cause the RestTemplate that wraps it to drop down to attempt
to extract a response body, even if it has aleady been done (to
construct the OAuth2Exception from the resource server), leading
to a new RestClientException that essentially masks the original
problem. The solution is to delegate to the DefaultResponseErrorHandler
which always throws an exception as the last resort (instead of
simply doing nothing).