Fix potential duplicate token problem in JdbcTokenStore
It shouldn't happen in practice, but the JdbcTokenStore, since it
didn't check for the existence of a token with the same value
before doing an insert, might break confusingly later. I think
I would recommend a uniqueness constraint in the database anyway
but now we have a "last one wins" rule.