Make access token setter for value more conventional
i.e. it has a void return type. Previously it had returned a
new object, in an attempt to normalize the hashCode and equals
implementations. I believe those may now occasionally be invalid
(e.g. if someone sets the expiration without setting the value)
in some circumstances, but the value of a token is still used
as the sole source of data for hashing.

Fixes gh-175