Extract ApprovalStore from AuthorizationEndpoint into UserApprovalHandler
The presence of the ApprovalStore in AuthorizationEndpoint was awkward and it
was only ever going to be needed if the UserApprovalHandler used it as well.
The best solution seems to be to add another method to UserApprovalHandler
to generate the model for asking for user approval (it's definitely in the
same area of responsibility).