Deprecate ImplicitGrantService and add new ImplicitTokenRequest
The ImplicitGrantService introduced a race condition in the implicit grant
(but only for concurrent requests from the same client with the same parameters,
so unlikely to be a problem in a real system). The best way to solve it seems
to be to deprecate that service (which is only used in the scope of a single
method in AuthorizationEndpoint, so there is no need for peristence or anything)
and replace it with a new way to carry the OAuth2Request down into the TokenGranter.

Fixes gh-212