Add getGrantType() to OAuth2Request
The OAuth2Request (and hence OAuth2Authentication) can now be
queried explicitly to find the grant type for the associated
token. If the token is being refreshed the grant type in the
OAuth2Request presented to the TokenEnhancer is the
original grant type, not "refresh_token".

Fixes gh-199, fixes gh-218