Add request parameters as "details" in password grant
The AuthenticationManager in the password grant now has access
to the request parameters (via Authentication.getDetails()). It still
has to support UsernamePasswordAuthenticationToken, so no existing code
requires changes. Note that the request parameters are available in the
OAuth2AuthenticationToken anyway, so the AuthenticationManager does *not*
need to copy the details to a successfully authenticated result (although
a ProviderManager will already do that by default).

Fixes gh-284