Save Authentication in case of changes after checking existing token
When the DefaultTokenServices are used to create an AccessToken from
a Authentication (e.g. in a token grant) and the AccessToken already
exists, it is looked up based on (typically) a subset of its full
properties (e.g. userid and client id and scopes). Since the other
properties might have changed since the token was created, it makes
sense to ask the TokenStore to store the new Authentication.

Fixes gh-283