Remove the @Bean TokenStore from default configuration
Previously the AuthorizationServerConfiguration had exposed a
TokenStore as a @Bean, mainly as a convenience for a
ResourceServerConfiguration that happened to be in the same context.
The problem is that there's a potential cycle if the user defines
their own @Bean TokenStore, and uses it to configure a
Resource Server, when it also happens to be part of an Authorization
Server.

Fixes gh-338