Remove builder convenience method for ClientDetailsService
The AuthorizationServerEndpointsConfigurer had a clientDetails() method
that was supposed to be for internal use (but unfortunately had to
be public so it can be called by a class in another package). Without
changing the packaging, the safest change at this point (even though
it's a public API) is to change the method signature and document it
to make it clear that it's not public.

Fixes gh-336