Add "stateless" flag to resource server configuration
Default to true, so that security context is cleared if no access token
is presented. Useful default in mixed servers where some resources are
not stateless, or where they might accept anonymous authentication.

Using default=true changes the behaviour of existing applications, but
in a way that makes them more secure (it's better to opt into anything
with a risk of authenticating accidentally).

Fixes gh-360