Add token type to OAuth2AuthenticationDetails
Clients that rely on the OAuth2AuthenticationDetails to figure out the
way to send authentication to a server could not, before this change,
take into account the token type to use in the header. Defaulting to
"bearer" sometimes fails (e.g. apparently on Facebook). There are no
such clients in Spring OAuth as it happens, but other projects are using
this feature so now we make the data available.

It's a little bit messy in the OAuth2ClientAuthenticationProcessingFilter
since the TokenExtractor implementation we have has protected methods
whose signatures do not permit the token type to be determined. Maybe
in 2.1 we can break those methods and make their function clearer.

Fixes gh-354