Add optional UserDetailsService to DefaultUserAuthenticationConverter
Now authorities (and other interesting details) can be stored externally
to the token, at least optionally. Often the token provider might have a
different idea of user authorities than the local service decoding the
token, so it helps to be able to extend the potential source of that
data.

The Principal in the resulting Authentication is the UserDetails, so
there might also be other interesting data in there for access
decisions.

Fixes gh-358