Clarify CSRF and session creation policy for resource server
It *is* possible to disable CSRF on a per-HttpSecurity basis, and
also the same for session creation policy (I believe). So rather
than making CRSF mandatory by default it is better to switch
it off and also set to SessionCreationPolicy.STATELESS.

N.B. if CRSF is *on*, even SessionCreationPolicy.STATELESS does
not prevent the creation of a session (since the CsrfFilter itself
has to look in the session for a token).

Fixes gh-339