Add authentication event publisher to OAuth2AuthenticationProcessingFilter
Authentication events are not being published (before this change) because they generally
only come from a ProviderManager (and there isn't one). The
OAuth2AuthenticationProcessingFilter seems like the most logical place to put this feature
(rather than the OAuth2AuthenticationManager, just because we can handle all failures in
the same place with the filter).

Fixes gh-393