Add JsonParser utility interface so that JWT can choose one
At runtime Jackson (1) might not be available on the classpath and up
to now the JwtAccessTokenConverter relied on it coming from
spring-security-jwt transitively. The latest version of that library
no longer requires an external JSON parser, so we need a way to
choose a converter/parser at runtime.

Fixes gh-391