Use custom error handler if status doesn't match
The basic motivation for this is that OAuth2RestTemplate should throw
HttpClientException more often (so it behaves more like a regular
RestTemplate). There are some OAuth2 specific responses that can be handled
internally, but by and large those are already handled by the
token granters and the client context filter.