Compare AOP targets instead of proxy to resolve token services
The logic to detect an existing default ResourceServerTokenServices
was flawed in the case that the two default ones were proxied (e.g.
as is the case with @EnableGlobalMethodSecurity. It has proved
rather difficult to find a test that breaks on this one, but the
sample project from gh-342 is fixed with this change.

Fixes gh-342