Switch off Boot default authentication manager
To fix gh-401 we added some smart stuff to the configuration of a
DefaultTokenServices so that in a refresh grant there is a
UserDetailsService available if one has been configured globally.
Unfortunately when Spring Boot autconfig is on, it creates one
that seems to take precedence over one created by the user. To
switch that off we need to add a @Bean of type AuthenticationManager.