Simplify addition of custom token granter
It wasn't hard in principle to add a custom granter, but since all
practical examples of granters need at least an
`AuthorizationServerTokenServices` and probably also a
`ClientDetailsService` it was quite inconvenient to do, because those
things are created for you by the same
`AuthorizationServerEndpointsConfigurer` that is used to override
the `TokenGranter`. (I.e. you would have to customize all 3 really.)

This change makes the common granter helpers (token services, client
details service, and outh2 request factory) available as lazy proxies
so even if the token granter is created before the configuration is
executed for those services it can still use them later. The convenient
`AbstractTokenGranter` can therefore bew used as a base class for instance.

There is a sample in test/annotation/custom-grant.

Fixes gh-422