Fix for inconsistency in the auto approval of scopes
Made the implementation of ClientDetails - e.g. the BaseClientDetails - solely
responsible for the decision if a scope can be auto approved. Marking a
ClientDetails - when using JdbcClientDetailsService - with 'true' for the
autoapprove value will still cause all scopes to be auto approved. The consent
screen will be skipped in this scenario.

Fixes gh-479, fixes gh-482