Fix for lossing authentication details on token refresh
This fix is for the case when user implements custom
AccessTokenConverter that puts extra information into the token in
convertAccessToken and extracts it into Authentication inside
extractAuthentication.

Those details were lost on the refresh token code path.

The fix makes DefaultTokenServices to preserve the details much
like it is already doing so when user has supplied custom
AuthenticationManager.

The fix also changes createRefreshedAuthentication access level to
protected so that user can modify scope narrowing logic. That is
to conform to the access level of other similar methods, such as
isExpired.