Support for other response types then code & token
Allow for injection of ResponseTypesHandler in the
AuthorizationEndpoint to support for other response types other then
the default 'code' and 'token'. The default implementation when
no override is configured handles 'code' and 'token'.

The AuthorizationEndpoint has been split up in a ResponseTypesHandler
and AuthorizationRequestViewResolver to prevent code duplications.

Both the AuthorizationEndpoint and ResponseTypesHandler use the
AuthorizationRequestViewResolver. The AuthorizationRequestViewResolver
can also be configured to override default behaviour.

Both the ResponseTypesHandler and AuthorizationRequestViewResolver can
be configured using the AuthorizationServerEndpointsConfigurer or
using xml style with a custom-response-types-handler-ref and
authorization-request-view-resolver-ref on the
oauth2:authorization-server

This commit / pull-request replaces
https://github.com/spring-projects/spring-security-oauth/pull/627

Resolves #626