Use thread context class loader to load classes during deserialization
Previously, SerializationUtils used a standard ObjectInputStream for
deserialization. For this to work, the type being deserialized had to
be available to SerializationUtils' ClassLoader. This doesn't hold
true in environments where there are multiple ClassLoaders involved
such as a servlet container with Spring Security OAuth installed as
a shared library, or Spring Boot's DevTools (see
spring-projects/spring-boot#5071).

This commit updates SerializationUtils to use Spring Framework's
ConfigurableObjectInputStream, configured with the current thread's
context class loader.