Publish authentication events in OAuth2 client authentication filter
The filter processing authentications but doesn't publish regular
AuthenticationSuccessEvent (or failure events). It only ever
publishes the (unrelated) InteractiveAuthenticationSuccessEvent
which is not a subclass, and is used in a different way by
consumers. Normally such events are published by an
AuthenticationManager but there isn't one in this filter.

Fixes gh-744