Remove global anoymous authenticator from resource server config
It is quite difficult (if not impossible) to configure a resource
server to authenticate some of its resources not using OAuth2
without this change. The global authentication manager should never
be anonymous generally, so it's better not to provide one at all
than try and guess what the default should be.

Fixes gh-737