Remove Embedded Redis
Embedded Redis is unreliable to start in various OS's. Specifically, it
does not start successfully in the CI environment.

This commit removes embedded Redis and requires the user to be running
Redis to run the tests. In order to support running the tests on a Redis
instance that is shared between tests we also ensure we use unique ids
for tests.

Issues gh-748