Add support for single scope String type
DefaultAccessTokenConverter originally expected a String Collection
for the 'scope' attribute in the raw token while
extracting it into an Authentication or AccessToken.
This commit adds support for a single String value for the 'scope' attribute
as certain Authorization Servers are returning a String type instead of a Collection/Array type.

Fixes gh-745