fix-JWT-refresh-tokens-unable-to-refresh-themselves
This change is needed because JwtTokenStore has no implementation for storeAccessToken and storeRefreshToken.
Which is fine, because JWT doesn't require persistence. But a refresh token is valid if
1) Digital signature is verified
2) There's an approval in the DB for all the scopes and the client_id, stored in the JWT itself.

When you refresh a token, the DefaultTokenServices removes all refresh tokens and access tokens. The only reasonable
way to 'remove' a JWT is to remove it's approval in the approval store. Since once removed it's never inserted again (user has to grant an approval again)
the next time you use a refresh token it will not be validated because of 2) and will fail.
A simple fix for that is providing an implementation for store(Access/Refresh)Token which just inserts an approval to the DB.