Add Elliptic Curve verification support for spring-security-jwt
The implementation is inspired by the RSAVerifier and uses code from org.apache.xml.security.algorithms.implementations.SignatureECDSA to transcode from JWK to JCE compatible signatures.

It is meant to be used by the patch to `spring-security-oauth2` which introduces support for *Elliptic Curve* based *JWK* keys.

This PR is one of the two PRs to deliver #1159