Add support for Elliptic Curve JWK signatures
This will utilize the changes to spring-security-jwt for handling EC signatures. It integrates fully into the existing codebase and will simply allow to verify signatures based on EC when they are read from a JWK definition.

This PR depends on #1161 to be merged. Once this PR is merged a new version of `spring-security-jwt` should be released and this PR needs to be updated to use the final version instead of the `1.0.9.BUILD-SNAPSHOT` version it currently uses.

This PR is the seconds part of delivering #1159