Ignore unsupported JWK types when parsing JWK set
At present, `JwkSetConverter` throws a `JwkException` when it attempts to parse a JWK with unsupported key type. This causes problems with providers that support multiple JWS algorithms, and therefore publish JWK set with keys of different types.

This commit updates the behavior of `JwkSetConverter` to ignore JWKs with unsupported key type.

Closes gh-1213