UTF-8 for oauth2ExceptionSerializer
Had an issue with modified exception, which contained UTF-8 symbols like 'š'. It was serialized to JSON as &scaron;. Generally, it's ok, but in my case I had to write my own serializer to accomplish the task of returning a proper UTF-8 exception message. Wouldn't it be better to use UTF-8 charset here instead of default ISO-8859-1?