Fix NPE when refresh_token is not provided
The NullPointerException can occur when the grant_type is refresh_token
but the refresh_token itself is not provided. And the response code
would be 500 as it is a server error.

With this fix, it will return 4xx as it is one of the client side error
that missed providing required parameter, refresh_token.

Fixes gh-941