Preserve user authentication details on re-authentication
During a token refresh in the `DefaultTokenServices` the user
authentication will be re-authenticated if an `AuthenticationManager`
was provided. A `PreAuthenticatedAuthenticationToken` is created
based on the user authentication and then passed to the
`AuthenticationManager`. However, if there were any details on
the user authentication those details are lost because they are
not copied to the `PreAuthenticatedAuthenticationToken`. If the
`AuthenticationManager` is not provided then this logic is skipped
over and the details are correctly preserved.

The fix is simply to set the details on the
`PreAuthenticatedAuthenticationToken` before passing it to the
`AuthenticationManager`.

Finally, I added two new tests to `DefaultTokenServicesTests` to
validate that the user authentication is built correctly on a refresh.
There is one test for the scenario when there is no re-authentication
which passes even before these changes and then the other tests the
re-authentication scenario which requires this change to pass.

Fixes gh-823